openapi: Move openapi.py into zerver/openapi.py.
Fixes #14006
